Refactor master.py flow, add run_full() function
Not tested yet, but in theory everything should now be ran from
master.py, and the logic has been implemented for appending unsent data
to a queue file. The unique identifier for each script was chosen to be
simply 'id(script)', which is foolproof during a single python runtime.
THIS NEEDS FURTHER INVESTIGATION, because I'm not sure if the id stays
the same across different runtimes, and if it didn't then terminating
master.py would automatically make the queue file (if it were to exist)
unapplicable for further use.

Code for reading the queue file is not written yet.

NONE OF THIS WAS TESTED, I'm short on time, kms.