Componetize master.py into external classes
master.py was getting too bloated, so I refactored the code into:

1. loghandler.py
2. queuehandler.py
3. filehandler.py

For now, I've written the general layout of each file using the pass
statement, but the idea is for master.py to only call existing methods
from each of those modules, as opposed to having built-in implementation
of each of them, which is messy and hard to read and comment.

All the code is still in master.py, it will be moved and reorganized in
a later commit.